+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Added `LocalizationUtility` (Draft)
 * Added `SysLanguageUidViewHelper`
 * Moved `LanguageViewHelper` under `Page` namespace